Use authenticated RPC proxy (#364)
### What

This replaces the usage of `RPC_URL` with the SDP proxy, allowing the
frontend to connect to authenticated RPC providers without exposing the
credentials in the frontend. Note that RPC is required by both the
dashboard and the upcoming embedded wallets and relies on different
authentication mechanisms; therefore, the endpoints are separated into
`/rpc/user` and `rpc/wallet` respectively.

### Why

Use RPC proxy instead of calling RPC directly.

### Known limitations

N/A

### Checklist

- [x] Title follows `SDP-1234: Add new feature` or `Chore: Refactor
package xyz` format. The Jira ticket code was included if available.
- [x] PR has a focused scope and doesn't mix features with refactoring
- [ ] `CHANGELOG.md` is updated (if applicable)
- [ ] Preview deployment works as expected
- [ ] CONFIG/SECRETS changes are updated in helmcharts and deployments
(if applicable)
- [ ] Ready for production